requirements.txt
asgiref==3.5.2
Django==4.1.3
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mysqlclient==2.1.1
Pillow==9.3.0
PyJWT==2.6.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
whitenoise==6.2.0
